orchestrator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
requests==2.32.3
pyyaml==6.0.1
